requirements.txt
# Flask Libraries
Werkzeug ~= 0.10
Flask ~= 0.10.1
webargs ~= 0.16
Flask-SQLAlchemy ~= 2.0
Flask-Login
Flask-Principal

## Database
SQLAlchemy ~= 1.0.4
sqlalchemy_utils ~= 0.30.2
psycopg2 ~= 2.6
Flask-Migrate ~= 1.4

# Utilities
Flask-Script ~= 2.0.5

# Chess Library
git+git://github.com/theovoss/Chess.git

# time
arrow ~= 0.5.4

# deployment
gunicorn ~= 19.3.0

# Testing
factory-boy
pytest
